Added two examples: p54.rs which ranks poker hands, and an example of using pattern matching as a conditional.
p54.rs is problem 54 from project euler.  it reads poker.txt (1000 2-player hands of poker) and returns
statistics of how many hands each player won and how man of each rank of hand appeared in the games.
pattern_match_as_cond_guard.rs is an example code showing a trick I thought up using pattern matching as
the guard on an if/else if/else structure for a sequence of operations that we want to have both return some
data and break from the if/else structure if successful