Added the C glue for the enttec and made many updates to EnttecPro.rs
Created the C file that wraps the functionality we can't easily import into Rust.
Created the necessary bindings to C and created a "safe" interface.
Implemented a phantom type to handle termios in Rust, and attempted to expose
a memory-safe interface to them by wrapping them in a struct and using its
constructor and destructor to handle memory allocation and deallocation on the C side.